Update Pre and Post Formation Compressions
This commit updates the previous Spherical Compression to use both Moment Purity Compression and Spherical Compression developed by Kram. 
Now AgX forms the image in P3, and then it gets compressed down to sRGB for the other two displays. `Moment Spherical Compression` is for pre-formation, `Moment Compression` is for post-formation.